Add files via upload
This folder contains files used to 'style' LaTeX documents as well as a working .tex file with the Capstone Project paper content titled CapstonePaper_Content.tex. This document should be updated as changes to the paper are made. This paper currently contains a bottom section of examples that should be removed before any submission. These files were provided by Springer to aid in the formatting of papers.